Minor refactoring to make importable package
