+++ src/dxss/solve_2d.py
@@ -12,10 +12,10 @@ from dolfinx.mesh import create_unit_interval
 from dolfinx.io import XDMFFile
 import sys
 sys.setrecursionlimit(10**6)
-from GMREs import GMRes
-from space_time import * 
-from precomp_time_int import theta_ref, d_theta_ref 
-from meshes import get_mesh_hierarchy, get_mesh_data_all_around
+from dxss.gmres import GMRes
+from dxss.space_time import * 
+from dxss.precomp_time_int import theta_ref, d_theta_ref 
+from dxss.meshes import get_mesh_hierarchy, get_mesh_data_all_around
 import pypardiso
 import scipy.sparse as sp
 import time
