+++ tests/slab_solver_test_3d.py
@@ -9,10 +9,10 @@ from dolfinx.mesh import create_box,CellType, GhostMode
 from dolfinx.io import XDMFFile
 import sys
 sys.setrecursionlimit(10**6)
-from GMREs import GMRes
-from space_time import * 
-from precomp_time_int import theta_ref, d_theta_ref 
-from meshes import get_mesh_hierarchy, get_mesh_data_all_around
+from dxss.gmres import GMRes
+from dxss.space_time import * 
+from dxss.precomp_time_int import theta_ref, d_theta_ref 
+from dxss.meshes import get_mesh_hierarchy, get_mesh_data_all_around
 import time
 import pypardiso
 import scipy.sparse as sp
